More robust plugin I/O mapping.
This does not change the actual mapping logic, but makes the application of
the mapping much more robust.  If there is no valid mapping for a given port,
that port is connected to silence (instead of crashing messily and/or via a
failed assertion).

Also tolerate mappings that nonsensically map to a buffer that is not present
(this particularly happens for MIDI ports in some cases) as a temporary fix.
The mapping logic needs work and/or our concept of just how much MIDI we support
in a route needs simplification...


git-svn-id: http://subversion.ardour.org/svn/ardour2/branches/3.0@10262 d708f5d6-7413-0410-9779-e7cbd77b26cf