Save references to external files in LV2 state as symbolic links.
With this commit it is possible to archive sessions that contain plugins with
complex file-containing state (e.g. Linuxsampler with a sample bank loaded)
with any archive tool, e.g.:

tar -hjcf session.tar.bz2 session

Which will give you a fully self-contained archive of everything used in the
session, so you can e.g. send it to a collaborator who may not have the same
sample banks in the same place as you and it will work.


git-svn-id: http://subversion.ardour.org/svn/ardour2/branches/3.0@10817 d708f5d6-7413-0410-9779-e7cbd77b26cf