Draw all automation points, rather than trying to use a heuristic to show some at any given zoom level.
This works because (1) automation data is obtained by sampling, and we control
the sampling interval (2) automation data is filtered to remove co-linear
points before becoming part of a ControlList. 

This commit removes some of the hairiest code in Ardour's GUI, which was
responsible for figuring out which invisible control points were affected by an
edit. The change is based on an experiment in mixbus done by Ben Loftis.


git-svn-id: http://subversion.ardour.org/svn/ardour2/branches/3.0@11798 d708f5d6-7413-0410-9779-e7cbd77b26cf