LV2: amend previous commit, respect replicated plugin instances
When a plugin is replicated (eg. mono plugin on a stereo-track),
configure_io() is only called for the first instance.
The 2nd instance will not have dedicated buffers but re-use copies
of the first instance. Should the 2nd instance use a LV2 worker-thread or
local communication ports they can not be shared with the first instance.
-> this patch allocates Atom-buffers for each instance.

PS. this still needs more work; there is only one UI instance for
all plugin instances. messages from the UI should go to all
instances.

git-svn-id: http://subversion.ardour.org/svn/ardour2/branches/3.0@13148 d708f5d6-7413-0410-9779-e7cbd77b26cf